Updated Date objects to LocalDateTime objects, refactored ids to longs, updated DB tables to reflect this, added testing specific DB capabilities, updated DAO queries, updated unit tests, added unit test specific SqlSessionFactory singleton

+++ backend/src/main/java/org/carpconn/dao/MangaDAO.java
@@ -12,13 +12,13 @@ import java.util.List;
  */
 public interface MangaDAO {
 
-    Integer create(@Param("manga") Manga manga);
+    Long create(@Param("manga") Manga manga);
 
-    Manga findManga(@Param("mangaId") int mangaId);
+    Manga findManga(@Param("mangaId") Long mangaId);
 
     void update(@Param("manga") Manga manga);
 
-    void delete(@Param("mangaId") int mangaId);
+    void delete(@Param("mangaId") Long mangaId);
 
     List<Manga> findAll();
